Minor Tweaks for Dashboard Type
